refactor: private/public 분기 처리
