+++ next-app/src/components/views/UserPage/List/ChannelList/ChannelList.tsx
@@ -1,18 +1,20 @@
 import { useRouter } from 'next/router'
 
 import List from '..'
-import usePostList from './hooks/usePostList'
+import useChannelList from './hooks/useChannelList'
 import { ITEMS_PER_PAGE } from 'components/views/MyPost/PostContainer.const'
 import Flex from 'components/common/Flex'
 import Paging from 'components/common/Paging'
-import EmptyContents from 'components/common/EmptyContents'
-import { SkeletonCardType } from 'components/common/Skeleton'
+import { SkeletonSubscriptionType } from 'components/common/Skeleton'
 import FeedItem from 'components/common/FeedItem'
-import type { PrivateItem } from 'types/feeds'
+import EmptyContents from 'components/common/EmptyContents'
+import type { PrivateSubscription } from 'types/subscriptions'
+import { useCheckIsMyProfile } from 'features/user/useCheckIsMyProfile'
 
-const PostList = () => {
+const ChannelList = () => {
   const router = useRouter()
-  const { listData, isLoading, isEmptyList, totalCount } = usePostList()
+  const { listData, isLoading, isEmptyList, totalCount } = useChannelList()
+  const isMyProfile = useCheckIsMyProfile()
 
   const totalPage = totalCount ? Math.ceil(totalCount / ITEMS_PER_PAGE) : 1
 
@@ -22,18 +24,18 @@ const PostList = () => {
         renderList={() =>
           isLoading
             ? Array.from({ length: ITEMS_PER_PAGE }).map((_, idx) => {
-                return <SkeletonCardType key={idx} />
+                return <SkeletonSubscriptionType key={idx} />
               })
             : listData?.map((item) => (
                 <FeedItem
                   key={item.id}
-                  type="card/private"
-                  item={item as PrivateItem}
+                  type={isMyProfile ? 'subscription/private' : 'subscription'}
+                  item={item as PrivateSubscription}
                 />
               ))
         }
         renderEmptyContent={() =>
-          isEmptyList && <EmptyContents content="저장한 게시물이 없습니다" />
+          isEmptyList && <EmptyContents content="구독 중인 채널이 없습니다" />
         }
       />
       <Flex style={{ width: '100%', padding: '44px 0' }} justify="center">
@@ -55,4 +57,4 @@ const PostList = () => {
   )
 }
 
-export default PostList
+export default ChannelList
