+++ next-app/src/components/views/UserPage/List/PostList/PostList.tsx
@@ -1,18 +1,20 @@
 import { useRouter } from 'next/router'
 
 import List from '..'
-import useChannelList from './hooks/useChannelList'
+import usePostList from './hooks/usePostList'
 import { ITEMS_PER_PAGE } from 'components/views/MyPost/PostContainer.const'
 import Flex from 'components/common/Flex'
 import Paging from 'components/common/Paging'
-import { SkeletonSubscriptionType } from 'components/common/Skeleton'
-import FeedItem from 'components/common/FeedItem'
 import EmptyContents from 'components/common/EmptyContents'
-import type { PrivateSubscription } from 'types/subscriptions'
+import { SkeletonCardType } from 'components/common/Skeleton'
+import FeedItem from 'components/common/FeedItem'
+import type { PrivateItem } from 'types/feeds'
+import { useCheckIsMyProfile } from 'features/user/useCheckIsMyProfile'
 
-const ChannelList = () => {
+const PostList = () => {
   const router = useRouter()
-  const { listData, isLoading, isEmptyList, totalCount } = useChannelList()
+  const { listData, isLoading, isEmptyList, totalCount } = usePostList()
+  const isMyProfile = useCheckIsMyProfile()
 
   const totalPage = totalCount ? Math.ceil(totalCount / ITEMS_PER_PAGE) : 1
 
@@ -22,18 +24,18 @@ const ChannelList = () => {
         renderList={() =>
           isLoading
             ? Array.from({ length: ITEMS_PER_PAGE }).map((_, idx) => {
-                return <SkeletonSubscriptionType key={idx} />
+                return <SkeletonCardType key={idx} />
               })
             : listData?.map((item) => (
                 <FeedItem
                   key={item.id}
-                  type="subscription/private"
-                  item={item as PrivateSubscription}
+                  type={isMyProfile ? 'card/private' : 'card'}
+                  item={item as PrivateItem}
                 />
               ))
         }
         renderEmptyContent={() =>
-          isEmptyList && <EmptyContents content="구독 중인 채널이 없습니다" />
+          isEmptyList && <EmptyContents content="저장한 게시물이 없습니다" />
         }
       />
       <Flex style={{ width: '100%', padding: '44px 0' }} justify="center">
@@ -55,4 +57,4 @@ const ChannelList = () => {
   )
 }
 
-export default ChannelList
+export default PostList
