+++ next-app/src/components/views/UserPage/MyPage/MyPageContainer.tsx
@@ -3,8 +3,8 @@ import { useRouter } from 'next/router'
 
 import Flex from 'components/common/Flex'
 import Tab, { getSelectedTab } from 'components/common/Tab/Tab'
-import ChannelList from './List/ChannelList'
-import PostList from './List/PostList'
+import ChannelList from '../List/ChannelList'
+import PostList from '../List/PostList'
 import { getFeedoongUrl } from './MyPageContainer.utils'
 import { useGetUserProfile } from 'features/user/userProfile'
 import PageContainer from 'components/common/PageContainer'
