+++ next-app/src/pages/mypage/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import MyPageContainer from 'components/views/MyPage'
+import MyPageContainer from 'components/views/UserPage/MyPage'
 
 const MyPage = () => {
   return <MyPageContainer />
